+++ src/BrokerLib/sys/ScriptService.h
@@ -12,6 +12,8 @@
 
 #include <string_view>
 
+#include <sol/sol.hpp>
+
 namespace dcclite::broker
 {
 	class Project;
@@ -20,5 +22,14 @@ namespace dcclite::broker
 
 namespace dcclite::broker::ScriptService
 {
+	class IScriptSupport
+	{
+		public:
+			virtual void IScriptSupport_RegisterProxy(sol::table &table) = 0;
+
+			virtual void IScriptSupport_OnVMInit(sol::state &state) = 0;
+			virtual void IScriptSupport_OnVMFinalize(sol::state &state) = 0;
+	};	
+
 	extern void Start(Broker &broker, const Project &project);
 }